etl_template/pyproject.toml
[project]
name = "etl_templates"
version = "0.1.0"
requires-python = ">=3.11.0"
authors = [
    { name = "Mark Zwart", email = "m_zwart@douane.nl" },
    { name = "Youri Jongejan", email = "y.jongejan@douane.nl" },
    { name = "Jeroen Poll", email = "J.Poll@douane.nl" },
]
readme = "README.md"
		
dependencies = [
    "alabaster==1.0.0",
    "anyio==4.9.0",
    "asttokens==3.0.0",
    "azure-core==1.32.0",
    "azure-devops==7.1.0b4",
    "babel==2.17.0",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "colorama==0.4.6",
    "decorator==5.2.1",
    "docutils==0.21.2",
    "duckdb==1.1.3",
    "executing==2.2.0",
    "flet==0.27.4",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httpx==0.28.1",
    "idna==3.10",
    "igraph==0.11.8",
    "imagesize==1.4.1",
    "ipython==9.1.0",
    "ipython_pygments_lexers==1.1.1",
    "isodate==0.7.2",
    "jedi==0.19.2",
    "Jinja2==3.1.5",
    "jsonpickle==4.0.5",
    "MarkupSafe==3.0.2",
    "matplotlib-inline==0.1.7",
    "msrest==0.7.1",
    "networkx==3.4.2",
    "numpy==2.2.4",
    "mermaid-py==0.7.1",
    "msrest==0.7.1",
    "mkdocs==1.6.1",
    "mkdocs-autorefs==1.4.1",
    "mkdocs-get-deps==0.2.0",
    "mkdocs-material==9.6.10",
    "mkdocs-material-extensions==1.3.1",
    "mkdocs-mermaid2-plugin==1.2.1",
    "mkdocs-panzoom-plugin==0.1.8",
    "mkdocstrings==0.29.1",
    "mkdocstrings-python==1.16.8",
    "oauthlib==3.2.2",
    "packaging==24.2",
    "pandas==2.2.3",
    "parso==0.8.4",
    "prompt_toolkit==3.0.50",
    "pure_eval==0.2.3",
    "Polars==1.26.0",
    "pyfiglet==1.0.2",
    "Pygments==2.19.1",
    "python-dateutil==2.9.0.post0",
    "python-json-logger==3.2.1",
    "pytz==2025.2",
    "pyvis==0.3.2",
    "PyYAML==6.0.2",
    "repath==0.9.0",
    "requests==2.32.3",
    "requests-oauthlib==2.0.0",
    "roman-numerals-py==3.1.0",
    "setuptools==75.8.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "snowballstemmer==2.2.0",
    "Sphinx==8.2.3",
    "sphinx-rtd-theme==3.0.2",
    "sphinxcontrib-applehelp==2.0.0",
    "sphinxcontrib-devhelp==2.0.0",
    "sphinxcontrib-htmlhelp==2.1.0",
    "sphinxcontrib-jquery==4.1",
    "sphinxcontrib-jsmath==1.0.1",
    "sphinxcontrib-mermaid==1.0.0",
    "sphinxcontrib-qthelp==2.0.0",
    "sphinxcontrib-serializinghtml==2.0.0",
    "sqlparse==0.5.0",
    "stack-data==0.6.3",
    "texttable==1.7.0",
    "traitlets==5.14.3",
    "typing_extensions==4.13.0",
    "tzdata==2025.2",
    "urllib3==2.3.0",
    "wcwidth==0.2.13",
    "xmltodict==0.14.2",
    "fastexcel==0.13.0",
    "pyarrow==19.0.1"
]


[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
